Unify and simplify the resolution of relationships via directives
