+++ src/Schema/Directives/Fields/BelongsToManyDirective.php
@@ -5,7 +5,7 @@ namespace Nuwave\Lighthouse\Schema\Directives\Fields;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
 use Nuwave\Lighthouse\Support\Contracts\FieldManipulator;
 
-class BelongsToManyDirective extends MultipleRelationDirective implements FieldResolver, FieldManipulator
+class BelongsToManyDirective extends RelationDirective implements FieldResolver, FieldManipulator
 {
     /**
      * Name of the directive.
