+++ src/Schema/Directives/Fields/HasManyDirective.php
@@ -5,7 +5,7 @@ namespace Nuwave\Lighthouse\Schema\Directives\Fields;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
 use Nuwave\Lighthouse\Support\Contracts\FieldManipulator;
 
-class HasManyDirective extends MultipleRelationDirective implements FieldResolver, FieldManipulator
+class HasManyDirective extends RelationDirective implements FieldResolver, FieldManipulator
 {
     /**
      * Name of the directive.
